biomedical-backend/requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.57.0
h11==0.14.0
idna==3.10
jax==0.6.0
jaxlib==0.6.0
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
